data_collection/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1
selenium==4.16.0

# Data handling
pandas==2.2.0
numpy==1.26.3
duckdb==0.10.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
